+++ monticore-generator/src/main/resources/methods/SupplierSet.ftl
@@ -1,5 +1,5 @@
 <#-- (c) https://github.com/MontiCore/monticore -->
 ${tc.signature("attribute")}
 ${defineHookPoint("Setter:Before")}
-this.${attribute.getName()} = () -> ${attribute.getName()};
+this.${attribute.getName()} = new de.monticore.symboltable.__internal__Supplier<>(() -> ${attribute.getName()});
 ${defineHookPoint("Setter:After")}
