+++ monticore-generator/src/main/resources/methods/opt/SupplierSet4Opt.ftl
@@ -1,5 +1,5 @@
 <#-- (c) https://github.com/MontiCore/monticore -->
 ${tc.signature("attribute", "nativeAttributeName")}
 ${defineHookPoint("Setter:Before")}
-this.${attribute.getName()} = () -> Optional.ofNullable(${attribute.getName()});
+this.${attribute.getName()} = new de.monticore.symboltable.__internal__Supplier<>(() -> Optional.ofNullable(${attribute.getName()}));
 ${defineHookPoint("Setter:After")}
